// src/config/config.type.ts
export interface IAppConfig {
    TWITCH_BOT_USERNAME: string
    TWITCH_BOT_ACCESS_TOKEN: string
    TWITCH_BOT_CLIENT_ID: string

    TWITCH_CHANNEL_NAME: string
    TWITCH_ACCESS_TOKEN: string
    TWITCH_CLIENT_ID: string

    TWITCH_EVENT_REDEEM_TTS_REWARD_ID: string;
    TWITCH_USER_BLACKLIST: string[]

    CONDA_BAT_PATH: string
}